# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cartoon Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user/listUser": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Listar usuários",
                "description": "Retorna uma lista de usuários",
                "operationId": "43d9f3dfd61538ac277ed191284b8856",
                "responses": {
                    "200": {
                        "description": "Lista de usuários"
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            }
        },
        "/api/account/storage": {
            "post": {
                "tags": [
                    "Session"
                ],
                "summary": "Armazenar um novo usuário",
                "description": "Armazena um novo usuário no sistema.",
                "operationId": "2d66ce3ed7440c5bcae5483d532e7098",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "lastName": {
                                        "type": "string",
                                        "example": "Tereuba"
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "email@email.com"
                                    },
                                    "cpf": {
                                        "type": "string",
                                        "example": "999.888.777.66"
                                    },
                                    "dataBirth": {
                                        "type": "string",
                                        "example": "01-01-1685"
                                    },
                                    "cellphone": {
                                        "type": "string",
                                        "example": "(27)9999-7777"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "johndoe@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "secret"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user/show": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get user information",
                "operationId": "5b2f8d2790414ef59a6903edac64a1da",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "ID of the user to get",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns user information",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "59"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "John"
                                        },
                                        "lastName": {
                                            "type": "string",
                                            "example": "Test"
                                        },
                                        "cpf": {
                                            "type": "string",
                                            "example": "979.508.820-34"
                                        },
                                        "dataBirth": {
                                            "type": "string",
                                            "example": "1895-05-05"
                                        },
                                        "cellphone": {
                                            "type": "string",
                                            "example": "279999-9999"
                                        },
                                        "image": {
                                            "type": "string",
                                            "example": "d043b2065e9ecf6d152a878703b364cc.jpg"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "email@email.com"
                                        },
                                        "address": {
                                            "type": "array[]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "User not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/update": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Atualizar um usuário",
                "description": "Atualizar um novo usuário no sistema.",
                "operationId": "0cfc432c796157e79e418318091f4305",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "lastName": {
                                        "type": "string",
                                        "example": "Tereuba"
                                    },
                                    "cpf": {
                                        "type": "string",
                                        "example": "999.888.777.66"
                                    },
                                    "dataBirth": {
                                        "type": "string",
                                        "example": "01-01-1685"
                                    },
                                    "cellphone": {
                                        "type": "string",
                                        "example": "(27)9999-7777"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Atualizado criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Token sem autorização."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro ao atualizar o usuário",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erro de validação."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao atualizar o usuário",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erro interno no servidor."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/destroy": {
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Deletar um usuário",
                "description": "Atualizar um novo usuário no sistema.",
                "operationId": "bcfdd09e84d8e398501bbe9e2a9cbc16",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operation successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Token sem autorização."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro ao atualizar o usuário",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Usuario não encontrado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao atualizar o usuário",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erro interno no servidor."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "title": "Id",
                        "description": "Id",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "title": "Name",
                        "description": "Name",
                        "type": "string",
                        "format": "string"
                    },
                    "lastName": {
                        "title": "Last Name",
                        "description": "last Name",
                        "type": "string",
                        "format": "string"
                    },
                    "cpf": {
                        "title": "Cpf",
                        "description": "Cpf",
                        "type": "string",
                        "format": "string"
                    },
                    "dataBirth": {
                        "title": "Cpf",
                        "description": "Cpf",
                        "type": "string",
                        "format": "string"
                    },
                    "cellphone": {
                        "title": "Cellphone",
                        "description": "Number of cellphone",
                        "type": "string",
                        "format": "string"
                    },
                    "image": {
                        "title": "Image",
                        "description": "Photo Perfil",
                        "type": "string",
                        "format": "string"
                    },
                    "email": {
                        "title": "Email",
                        "description": "Your Email",
                        "type": "string",
                        "format": "file"
                    },
                    "password": {
                        "title": "Password",
                        "description": "Your Password",
                        "type": "string",
                        "format": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            }
        },
...
    }
}